+++ packages/mqtt-bridge/src/index.ts
@@ -1,0 +1,14 @@
+/**
+ * @license
+ * Copyright 2026 ThirdReality
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+/**
+ * @matter-server/mqtt-bridge - MQTT bridge (matter2mqtt) for matter.js server
+ */
+
+export * from "./MqttBridge.js";
+export * from "./MqttConnection.js";
+export * from "./OnOffState.js";
+export * from "./Topics.js";